Adds initial salty typescript app
Converts a php app to typescript for deployment on deno deploy. The original PHP application by neatnik uses sodium_crypto_secretbox for encryption, which is based on XSalsa20Poly1305. The Web Crypto API available in browsers and Deno does not natively support XSalsa20Poly1305. The closest secure alternative widely available is AES-GCM.

This means that the encrypted text generated by this new Deno TypeScript version will not be compatible with the encrypted text generated by the neatnik PHP application, and vice-versa. It is not possible to decrypt messages encrypted by the PHP app using the Deno app, or vice-versa.