style(ui): set max width to 65 char for top and modal p
When you set `max-w-[65ch]` alone, the p will left-align, because max-width
constrains the element's size, but it doesn't automatically center it.

To apply max-width and keep the paragraph centered, you need to combine it
with an auto margin on the left and right. In Tailwind CSS, this is achieved
with the mx-auto class.

Combining both, will make the para act like it's set with "prose". Set on the
top p under the title, and the modal help.